+++ src/main/webapp/resources/js/pages/UserGroupsPage/components/UserGroupDetailsPage.jsx
@@ -163,7 +163,14 @@ export default function UserGroupDetailsPage({ id }) {
           <UserGroupProjectsTable groupId={id} />
         </TabPane>
         {state.canManage ? (
-          <TabPane tab={i18n("UserGroupDetailsPage.tab.delete")} key="delete">
+          <TabPane
+            tab={
+              <span className="t-tab-delete">
+                {i18n("UserGroupDetailsPage.tab.delete")}
+              </span>
+            }
+            key="delete"
+          >
             <Title level={4}>
               {i18n("UserGroupsDetailsPage.title.delete")}
             </Title>