+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Changes
 * [UI/Developer]: Removed `irida.cart.js` (an older angularjs service for the cart).
 * [UI]: Fixed issue #637 where metadata toggles were not displaying in the "on" position properly on loading the advanced phylogenic viewer.
 * [UI]: Removed unused front end dependencies (`angular-boostrap-lightbox` and `lodash.keys`).
+* [Developer/UI]: Removed `irida` angular app.  Individual components load themselves now.
+* [Developer/UI]: Updated to Ant Design v4.4.2.
 
 20.01 to 20.05
 --------------